Update go main and frontend for webassembly interaction
Update go main and frontend for webassembly interaction:

  * create index.js according tinygo examples.
  * split main.go into main_tinygo.go and main_notinygo.go with build tags
    for compiling with and without tinygo.

main_tinygo.go defines the entrypoints to be used in the web-interface:
tokenize, parse, evaluate and expects the JS implementation to provide:
printJS.

main_notiny.go is currently aimed to be a CLI utility, to extended to
run a basic r[e]pl for tokenize, parse and evaluate. All these code
changes are still in the realm of bootstrapping.

In order to build and deploy things rework Makefile to be a bit more
readable and work with node firebase app.

Use linear memory to pass strings between JS and go/wasm:
Passing strings between JS and wasm is unexpectedly difficult. Go
provides syscall/js which would allow for reading/writing the Dom
directly. However, that results go code closely coupled to DOM and JS
and makes it unusable with the more abstract WASI (WebAssembly System
Interface). Additionally and more importantly, when reading the
DOM tinygo prints the following error:

  syscall/js.finalizeRef not implemented

which according to tinygo issue 1140 causes a memory leak. I settled for
passing strings by converting to byte arrays and writing them to wasm's
linear memory

Link: https://tinygo.org/docs/guides/webassembly/
Link: https://wasmbyexample.dev/examples/hello-world/hello-world.go.en-us.html
Link: tinygo-org/tinygo#1140
Link: https://www.wasm.builders/k33g_org/an-essay-on-the-bi-directional-exchange-of-strings-between-the-wasm-module-with-tinygo-and-nodejs-with-wasi-support-3i9h
Link: https://www.alcarney.me/blog/2020/passing-strings-between-tinygo-wasm/
Signed-off-by: Julia Ogris <julia.ogris@gmail.com>