parser: Refactor parser.go into more files
Refactor parser.go into:

- parser.go
- parser_literal.go
- parser_type.go

`parser.go` is becoming increasingly large and unwieldy. It's harder
to find ones way around and presumably this will become more of a
problem as the implementation becomes more complete. I'm happy to
revert this change is considered more confusing and then helpful.

There are no functional changes in this commit. It is made preparation
for implementing composite literal parsing.