parser: Add scoped typed analysis to declarations
Add scoped typed analysis to declarations, ensuring that variables or
parameters do not get redeclared in same block. Variables may be
re-declared in inner scope of a new block with a new type which does
not affect the outer scope. Add detailed tests for various scenarios.

As the parser will also take care of type analysis this is a necessary
addition and fixes a previous omission.