go: Refactor scope in parser and evaluator
Refactor scope in parser and evaluator for both to be private. Rename
`newEnclosedScope` to `newInnerScope` as this term confuses my little
brain a little less. Because `newInnerScope` is now private in
evaluator it triggered an "unused" linter warning, so remove it. We
will re-add it when we need.

It'd be neat to use generics with scope and then re-use the same
implementation for evaluator and parser. However, tinygo doesn't
support generics yet, so we have to do a little bit of copying here.