evaluator: Implement function evaluation
Implement function evaluation, including return values. For this we need
to implement block evaluation to be reused by if, for and other
statements. A block receives a new inner scope to populate which will
not be valid outside the block. In the function call case, the inner
block has the argument values pre-populated. Update parser AST to hold
function declaration with function call node for easier evaluation.