parser: Parse `if` statements
Parse `if` statements into If AST node. While at it, abstract the way we
create block statements with a configurable set of terminating tokens.
`else` for example is terminating for if statements, but not for
function bodies, else bodies or loop bodies.

Also update block statement parsing, to NOT parse past the END token so
we can read and validate the block end token from the call site.