github: Replace firebase-deployment action
Replace firebase-deployment action by hand-crafted make targets. This has
been motivated by an apparently broken firebase-tools release (4 hours
old) that cannot be down-graded when using the firebase GitHub action
and is a blocker. Aside: the foxygoat authors believe it is nice to
stay third-party GitHub action free where possible so that the bulk of
what's run on CI can also be run locally.