github: Add PR comment with deployment URL
Move firebase deployment and PR commenting machinery into separate
script as it is a bit verbose for a Makefile.

Post comment with new deployment URL derived from PR number and git sha
of deployed commit. Update comment if already posted previously.

The script can also be run locally and does not comment on a PR with
`make firebase-deploy` which deploys to to `dev` channel. On CI `make
firebase-deploy` the deploy scrip finds the active PR number and
deploys to channel `prNN`. To deploy to prod run `make
firebase-deploy-prod` which deploys to the live channel.

Signed-off-by: Julia Ogris <julia.ogris@gmail.com>