parser: Refactor alwaysReturns to alwaysTerminates
Refactor alwaysReturns to alwaysTerminates as a non-functional change
and pure rename. This is another preparatory step for implementing
break statements as breaks interrupt the flow like returns but
_are not_ returns.

It will be used as follows:

	while true
		if true
			return
		else
			break
		end
		print "unreachable"
	end

We can now use AlwaysTerminates instead of AlwaysReturns to also check
that a function with return specified type always returns, because loop
statements _never_ always terminated, as the loop body could be
executed 0 times.