parser: Implement Pratt expression parser
Implement Top Down Operator Precedence or Pratt parser, replacing the
existing but very limited parseExpression method.

This parser is based on Thorston Ball's Interpreter book and his monkey
interpreter.

Functional changes include some error messages and the removal of type
prefix for array and map literals. The type is now inferred:

	arr := [1] // num array; arr := num[1] is now invalid

All arithmetic and logical expressions are now parsed. Outstanding work
includes array slices, expression parsing for the left hand side of an
assignment, e.g.

	arr[0] = 2

The evaluator implementation is also not included in this commit as it
has become fairly large as is.

Documentation updates are in flight in parallel at PR 31

Link: https://github.com/juliaogris/monkey/blob/master/parser/parser.go