evaluator: Rework scoped Value updates
Rework scoped Value updates such that instead of replacing the value in
the scope on assignment with scope.set(name, newValue), we now update
the Value field with:

	val := scope.get(name)
	val.set(newVal)

This is a preparatory step for index and dot expressions which also
update the "internals" of a Value.

As a fall-out of this step we had to extend the Value interface with a
`Set(Val)` method. We also had to remove the bool constants TRUE and
FALSE, as the internals of these constants got updated. `scope.getScope
(name)` is now not needed any longer and removed. Additionally
introduce an `Any` type that implements Value and wraps a value so that
we can update an any value from one type to another, e.g. Bool to
String.