evaluator: Change Array.Elements and Map.Order to pointer
Change Array.Elements and Map.Order to pointer. This is another
preparatory step for proper array and map manipulation so that two
variables may reference the same array or map and if one gets updated
the other one will get updated too.

In go maps are references and natively support this property, however
slices may get allocated new memory and are not guaranteed to keep the
same reference and appended.