parser: Add for ... range statement parsing
Add for statement to parser which has to match: `for IDENT := range
RANGE_EXPR`. Range Expression can be an expression evaluating to a
string, a map or an array or it can consist of 1 to 3 numbers forming a
StepRange.

Add 2 new AST nodes: `For` and `StepRange`. Fix up previous for loop
tests that now fail because of unused variable errors.