parser: Fix double index in assignment
Fix double index in target of assignment statements:

	a := [[1 2] ([3])]
	a[0][0] = 1

previously caused a parse error because we stopped further evaluating on
the target side after the first dot or index expression, assuming that
the index/slice parsing took care of it itself, but that was delegated
to the higher order `parseExpr` which is too broad for the left side of
an assignment (the target).

Fix by looping until next token is not a `[` or `.` any longer.

Also add test to evaluator ensuring we are evaluating correctly.