parser: Refactor map and array of no type from EMPTY to GENERIC
Refactor map of no type from EMPTY_MAP to GENERIC_MAP and similarly
refactor array of no type from EMPTY_ARRAY to GENERIC_ARRAY.

We are currently using empty maps `{}` and arrays `[]` of no type as
literals in expressions and assignments and give them a type in a
deferred manner: e.g. [1 2] + [] makes the second array an array of num
only when needed.

We also now allow parameters to be of type GENERC_MAP or GENERIC_ARRAY,
which can can only be used in builtin functions captured in the changes
to the function `assertArgTypes` in parser.go

In a follow up commit we are going to reuse this concept not only as
literal for delayed type binding but also to mark a parameter type
as "generic" map or array. This means a parameter hast to a be map, but
it could be a map of any, num or other types.

This will be useful for discovering invalid types for builtin functions
`has` and `del` when the first parameter passed is not a map.