github: Fix app_token script
Fix `app_token` script by using `openssl enc -A -base64` instead of
`base64` because `base64` adds linebreaks on linux but not on mac. The
flag to change this doesn't work on mac.

Reinstate usage on with `make release` and remove node based third-party
app. We had only temporarily changed this to see if the rest of the
brew formula update process in repo foxygoat/homebrew-tap would work.