frontend: Refactor CSS
Refactor CSS, removing the abstract class `pane` and replacing it by
direct references to `code` and `out`. This is a preparatory step for
adding a <canvas> element for which we will need a quite a few more
separate CSS rules for the `out` class, so unifying `out` and `code`
seems to make things more complicated.

At this stage this change has no functional or visual consequences.