frontend: Fix indentation at the end of sample code
Fix indentation at the end of sample code so that when entering a
newline no whitespace follows. The HTML source now doesn't look
correctly indented around the textarea, but I suppose that's how it
goes. For now. Need to rework layout completely. This is just a quick
fix for a demo to the kids tomorrow.