frontend: Fix output height
Fix the output height. Previously it was about 3 lines high which made
it near unusable. Adding the height 100% made the drawing area
(arguably) too smaller, so with some trial and error we've fixed that
to 40vh (40% of the viewport height). Again this is messy and needs a
proper rework in the hopefully no too far future.