evaluator: Clean up Run function
Clean up Run function to replace RunWithBuiltins as this is the most
common use case. Update evaluator tests to use function that runs the
evaluator and captures the output which reduces duplication. This is a
simplification step towards having Run be replaced with Init() and
Next() so we can use evy in debug mode and be driven by event loop.