frontend: Launch wasm from go main()
Launch wasm from go `main()` rather than from go exported function
`evaluate()`. It turns out that some things, such as time.Sleep and
fmt.Sprintf only get properly intialised with the main function is the
entry point to the wasm code, so let's fix this up.

We cannot pass parameters to main() so we need JS to provide function to
extract the evy source as string. Unfortunately we can't pass strings
directly in wasm so we have write bytes to linear memory. This was fine
when passing strings as function arguments, but when we want return
values this falls apart as wasm can only provide a single return value.
I attempted to make this single return value a pointer in memory where
the first 4 bytes contain the length and the following bytes the
contents of a string, but got lost with endian specs and other details.
Instead JS provides the somewhat hacky functions `sourcePtr` and
`sourceLength`, which are called separate from Go code and reassembled
to a string.