all: Implement stop
Implement stop in frontend by turning the Run button into a Stop button
during execution. For this we also need an onExit callback into JS from
go when execution finishes so that the button label gets reset.

Additionally we need to occasionally yield to JS event processing via a
`time.Sleep` in the evaluator code. This operation is abstracted via
the Yielder interface.

Execution is stopped in Eval if the Stopped flag is set, which gets set
on click via the exported Go function stop(). The access to the stop
flag is not synchronized, but as JS is single threaded so we don't need
to worry about it.