frontend: Optionally load evy source via URL fragment
Optionally load evy source via URL fragment, where the fragment has to
be `#source=<SOURCE_URL>`, e.g.:

	https://evy-lang--73-51oq7tbx.web.app#source=samples/draw.evy
	https://evy-lang--73-51oq7tbx.web.app#source=https://gist.githubusercontent.com/juliaogris/ff91697865533c24d0ad52df24c4023e/raw/f42b0a58d693dd5249245e99aaefc11bf7494341/draw.evy

Of short fragment for files in samples/*.evy

	https://evy-lang--73-51oq7tbx.web.app#draw

Choose fragment rather than query parameter in URL so that `/` and `:`
don't have to be decoded. Source loading happens client side, so
arguably this is more correct anyway.

Add `hashchange` listener so that the source gets updated when the hash
gets updated, we should follow up with a drop-down for all the
samples.

While at it, add a `"keydown"` listener for the ctrl+enter, so we don't
need to fiddle with the run button any more.