all: Implement up, down, move and key events
Now that all prep work and refactoring has been split off, let's get to
the meat and implement up, down, move and key events.

There is a bit of complexity in translating pixels represented as
`pointerevent.offsetX` and `pointerevent.offsetY` into the correct evy
coordinates.

In the evaluator we have introduced an Event type the has a Name field
by which the evenHandler is looked up. It also has a Params fields
which contains []any to be converted into evaluator.Value and then
passed to the named eventHandler for execution. All this fell in place
pretty easily and had many similarities with function calls.

The last puzzle piece is wiring (and unwiring) the eventHandlers _after_
the main execution of the evy program is finished. For this we export
`registerEventHandler("down")` from JS which currently knows of down,
up move and key events. If the eventHandler gets registered, JS calls
out to wasm on the relevant event vial exported Go/Wasm function:

	onUp(x,y float64)   { /* add evaluator.Event to queue */ }
	onDown(x,y float64) { /* ... */}
	onMove(x,y float64) { /* ... */}
	onKey(s string)     { /* ... */}

An event loop in the Go/Wasm main.go file, in function handleEvents
() works all events of queue and sleeps when for a millisecond to give
JS events a chance to processed if queue is empty. If there aren't any
eventHandlers the evy program stops by itself, otherwise the stop
button needs to be pressed.

Test it with https://evy-lang--73-51oq7tbx.web.app/#source=samples/draw.evy

This was the last big chunk of unimplemented evy syntax (I think).