wasm: Rework yielding so that it doesn't interfere with sleep
Rework yielding (again 😿) so that is doesn't interfere with sleep.
With the currently deployed evy playground an endless loop without
a `sleep` crashes the browser tab, as the most recent Yielding bug-fix
has introduced a new bug. sigh.

In this fix we reset the yield timeout and eval counter every time we
run `sleep` as part of the evy program. This too seems fragile, but I'm
running out of ideas and empirically it seems to work better.