builtin: Add printf and sprintf functions
Add printf and sprintf functions using Go's fmt.Sprintf (which now that
we hoist wasm code via main works 🚀). To allow for formatted printed
of numbers and bools add and `unwrap` function to value.go, unwrapping
the value of the Value interface implementation for basic types only.

In the process of creating printf, we discovered some missing unescaping
in the lexing stage as the goal is to keep string handling consistent
with Go and print "\n" as newlines. Fix it and improve error message
for invalid strings.