wasm: Rework access to evy source code
Rework access to evy source code. What we wanted was a JS function that
returns a string (the evy source code). However, go/wasm only allows JS
functions that return float64, int, int32 or uint32 it seems. As a
result we queried the global sourcePointer and sourceLength as written
to linear memory separately and rebuilt the string in go wasm.

Now that we will need another string returning functions (read()) it is
time to rework this approach: Write the string to memory in JS all the
same, but combine pointer (high 32 bits) and length (low 32 bits) into
a single 64 bit JS Number which is return to in go/wasm as float64.
Reconstruct pointer and length from it in go/wasm with the inverse bit
manipulations and reconstruct string.

It could be that passing float64 is going to bite as at some point - we
will deal with it.