all: Add read function
Add blocking read function for line-by-line reading from text input.

Add jsRead function to index.js which returns a memory address of read
string iff a string has been entered into the input area followed by a
newline, otherwise return 0.

go/wasm provides read() which wraps jsRead() in a loop with intermittent
sleeps to yield to JS for processing user events. If the return value
of jsRead() is not 0 extract the string for the given memory address
and return. read() is provided to the evaluator via the runtime struct
and used inside the evy builtin read.

Move `hidden` class towards the end of index.css because for it so that
the same property `display` can be overriden from `flex` to `none`.