frontend: Hide and show evy widgets via URL fragment
Hide and show evy widgets via URL fragment so that we can hide sliders
by default but demonstrate their use in the follow up commits, e.g.:

	http://evy.dev/#show=.read,.slider&hide=.canvas,.output

Note that the value of `show` and `hide` is directly used as
querySelector. If elements are to be selected by className a leading
dot is required.