animate: Fix flicker by reworking yielding
Fix flicker for animations via updated yielding. The yielder yields
about every 100ms to check if the cancel button has been pressed
(in case of an accidental endless loop).
Add a Reset method to the Sleeping yielder so that that the counter be
reset on all external sleeps and before executing a an eventHandler.
Additionally use the yielding sleep for the reading in JS, so make
read a Method on the yielder.