parser: Allow for multiline array declarations
Allow for multiline array declarations so that it is easier to declare
large blocks of data. We want to get this in before we look into `evy
format` or syntax colouring as this might impact things.

Initially I tried making it part of `parseExprList` but it became a bit
messy - so sticking with "a little bit of duplication is better..." and
implementing a separate parseExprListWithNL instead. In a follow up
commit we will tackle multi-line map declarations.