go: Rename all AST nodes representing statements
Rename all AST nodes representing statements with the suffix Stmt. We
will want to to step / stop on a per statement basis when we implement
debugging, and in a follow up commit we will add comments to Statements
so that we can use the AST for formatting.

It will not be possible to attach comments to (sub-) expressions.