frontend: Fix read UI
Fix read UI, so that it does not keep on grabbing the focus, e.g.:

1. Run program with `s := read`
2. Focus gets set to read input field and reset every 50ms

It is impossible to edit the source while the program is running. Change
the code so that the focus is only set to the input text field only the
first time after clicking run. While at it also ensure that input field
is shown on the first read.