parser: Add formatting
Add formatting for all node types, except for multiline array and map
literals (they will be pulled together into a single line).

This is the second attempt, in a first attempt the node interface was
extended by a `Format(w io.StringWriter)` method which did the job but
spread the code all over the already big ast.go file. In this attempt
we keep almost all formatting concerns outside of the ast and collected
together in the type formatting and the file format.go.

We also attempted to have the EmptyStatement collect all consecutive
empty lines and and stand alone comment lines, which ended up less
readable than one EmptyComment ast node per line.

In the follow up commits we will address multi-line arrays and maps,
wiring with evy command and frontend and formatting all samples.