parser: Add formatting
Add formatting for all node types, except for multiline array and map
literals (they will be pulled together into a single line).

This is the third attempt, in a first attempt the node interface was
extended by a `Format(w io.StringWriter)` method which did the job but
spread the code all over the already big ast.go file. In a second
attempt we keep almost all formatting concerns outside of the ast and
collected together in the type formatting and the file format.go.
Unfortunately we tripped over a tinygo map issue doing this(see gophers
slack link below). In this third attempt we track comments and wss with
the node but still keep formatting in a separate file and data
structure.

We also attempted to have the EmptyStmt collect all consecutive empty
lines and and stand alone comment lines, which ended up less readable
than one EmptyStmt ast node per line.

In the follow up commits we will address multi-line arrays and maps,
wiring with evy command and frontend and formatting all samples.

Link: https://gophers.slack.com/archives/CDJD3SUP6/p1677288675055869