frontend: Wire formatting
Wire source code formatting with the frontend. We now format all source
before it is being executed in the web interface. Unfortunately we
can't pass os.Argv to wasm main so we indiscriminately parse format all
code before evaluation and don't allow for stand alone formatting
through the web interface (for now).

If the source code can be parsed successfully the formatted code is
written back to the input field and re-parsed for evaluation to avoid
incorrect error message offset.