parser: Refactor parser error handling
Refactor parser error handling by returning a slice of parser.Errors,
also implementing error, directly. This is a further step to make the
parser stateless and ultimately private. It allows us to introduce
the following top-level function in the parser package:

	ast, err := Parse(input, builtins)

In a follow up commit make the parser private.