frontend: Slide to code screen when navigating
Slide to code screen when navigating to new code sample on mobile.
Otherwise we might stay on the output screen, possible still with the
stopped output of the previous sample which is confusing.

Later we may introduce some #autorun which will keep things on mobile on
the output screen and run the new code.