frontend: Rework sample structure
Rework sample structure by renaming samples to courses. This aligns with
the UI implementation and the JSON content in former samples.json. This
restructure is happening in preparation for real course material
additions. Eventually we will probably separate playground with samples
from courses, but for the time being they have to do double duty.