frontend: Improve cmd+enter
Improve cmd+enter shortcut to start a program so that key events will be
listened to straight away. If the focus remains on the editor the key
events are ignored so that code can be edited while a program is
running. Before this change one had to click on the canvas before key
events came in to the evy program.