frontend: Lint CSS with stylelint
Lint CSS with stylelint, use --fix flag and manually fix 5 remaining
errors. We are not making stylelint part of the build process with
`make ci` yet, because it requires the setup of a node module :(.
We will follow up once we add bundling.

Style line was installed with:

        npm init #  yesyesyes
        npm init stylelint
        # add `"rules": { "rule-empty-line-before": "never" }` .stylelintrc.json
        npx stylelint --fix "**/*.css"

The only violated rule that couldn't be auto-fixed was
`no-descending-specificity`: "Disallow selectors of lower specificity
from coming after overriding selectors of higher specificity." Fair
enough this has bitten me before.

Link: https://stylelint.io/user-guide/rules/no-descending-specificity