frontend: Highlight lines with error
Highlight lines with evy errors in editor. This is done by finding the
correct token by its line and columns and adding the "err" CSS class to
it. It is becoming increasingly obvious that a more fully fledged web
code editor will be needed. Evy files with more than 700ish lines are
starting to get janky when editing. I could work around this by only
updating the relevant lines, but it all seems to get a bit to hacky.

After this error highlight addition we will leave editor hacking for
now.