parser: Add Token() to Node interface
Add Token() to ast Node interface so that we can more accurately
pinpoint the error position, e.g.:

	func numf n:num
	    print n
	end

	numf "abc"

The example above only highlighted the end of line with the `numf` with
a message saying first argument is of wrong type, but now it also
highlights `"abc"` correctly.

This is still problematic for error sources that are whole nodes, e.g.:

	numf "abc"+"123"

Should probably highlight all of `"abc"+"123"` rather than just the
first token. However, the might end up as undesired result for _very_
long expressions or multiline array / map literals.