frontend: Fix empty editor bug
Fix empty editor bug, which is caused by the editor module not
responding when its value/content changes to "". "" is falsy - oops.
This was the result of highlighting errors where we'd call

        editor.update({errorLines})

without content value. To keep the value for new value being undefined
we changed `this.value = value` to:

        this.value = value || this.value

This promptly didn't update for value === "" (falsy) and resulted in the
weird behaviour when emptying the editor. Fixed now with explicit check
again null and undefined.

This had to happen at some point. 🎉