frontend: Reset error highlight on start
Reset error highlight on start (run), so that the error highlight so
that we accurately display new errors or no errors if they are gone.
Previously errors only got reset if the line was updated which
incorrectly kept error line highlight for follow-on errors, e.g.:

	x := "foo"  // line 1: 'x' declared but not used
	print y     // line 2: unknown variable name 'y'

Fixed as

	x := "foo"  // 💣 error remained: line 1: 'x' declared but not used
	print x     // no error

Now all errors get reset.