💋 parser: Fix panic on nil token in error (#120)
Fix panic on nil token in error when overriding builtin function names.
This was noted when implementing `clear` which had a different
signature from the hand written clear in the samples. Add panic in
error appending for nil tokens so that we get traceable panics.

Pull-Request: https://github.com/foxygoat/evy/pull/120