🐭 frontend: Fix pointermove event for mobile (#126)
Fix pointermove event for mobile touch according to
https://stackoverflow.com/a/48254578

It seems that on mobile the browser captures pointermove events for its
own use (scrolling). This can be disabled with the CSS property
`touch-action: none;`. However, this is not preferable for
accessibility cases, so still allow pinch to zoom.

Tested on Android and iPad

The following sample now also works on mobile:
https://evy-lang--126-hesnlymh.web.app/#draw

Pull-Request: https://github.com/foxygoat/evy/pull/126