evaluator: Implement short-circuit evaluation of AND/OR
Implement short-circuiting of the AND and OR operators so that the
right-hand side is not evaluated if the result can be determined by the
left-hand side. For AND, this means left == false, since the whole
expression will be false regardless of what right is. For OR, this means
left == true, for the same reason.

Update the spec to say these operators perform short-circuit evaluation.