parser: Add type assertions
Add parsing and formatting of type assertions in expressions. As per the
spec, type assertions allow you to "cast" an `any` type to a concrete
type with the syntax `var.(type)` where `var` is an assignable of type
`any` and `type` is a concrete type. The type of the type assertion is
`type`.