eval: Add token to evaluator error
Add token with line and column location to evaluator errors so that we
can surface in UI _where_ a runtime error is happening. We are taking a
conservative approach here - some errors could be modelled as panics
and should never occur, much rather they should be parse errors.
However, because panics cannot be handled in TinyGo wasm at this stage
and therefore not be surfaced in UI we will go with standard go
errors.