✂️ main: Drop trailing newline in CLI read function (#134)
The CLI runtime `Read()` function was returning the line read with the
trailing newline. This is different to how the browser/wasm
implementation works which does not include the trailing newline. This
makes for different behaviour when running an evy program with the
different runtimes.

Pull-Request: https://github.com/foxygoat/evy/pull/134