samples: Add text-based tic tac toe
Add text-based tic tac toe game that can also be run in the terminal with

	evy run tictactoetext.evy

This is a precursor to the canvas / mouse event tictactoe as it allows
for more functional code with less global scope and less graphics and
drawing noise.

We don't want to list tictactoetext under the modal sample selection
menu, so we introduce the `"unlisted"` attribute in the sample.json. We
will still be able to link to it, it just doesn't show up in the menu.

We also had to increase the tinygo stacksize again because of the
expensive minmax calculations for a perfect computer player.

Unfortunately evy's performance is to slow to run minmax from the first
move, so use a random move instead.