eval: Remove ValueTypes BUILTIN, FUNCTION, BREAK, RETURN_VALUE
Remove the `BUILTIN` and `FUNCTION` `ValueType` values as they are not
used - `FUNCTION` is not used at all and `BUILTIN` is returned by a
single method that is not called.

Remove the `BREAK` and `RETURN_VALUE` `ValueType`, replacing the single
use of them with a type assertion.

These changes remove the evaluator type values that do not have a
corresponding parser type paving the way to replace `ValueType` with
`parser.Type`, enabling deeper type inspection at run time.