🛠️ eval: Fix break to not break more than one level (#138)
Fix the evaluation of the `break` statement so as to not break out of
more than one loop (`for` or `while`). We do this by turning a `Break`
`Value` into a `nil` `Value` when it is handled, breaking a loop.
Further outer loops will no longer break out.

Add tests for this case.

Pull-Request: https://github.com/foxygoat/evy/pull/138