eval: Replace ValueType with *parser.Type
Remove the `ValueType` type in favour of `*parser.Type` to enable deeper
type knowledge of composite types (arrays, maps) that do not have
values. The `ValueType` type only modelled an array or map but not the
type of its contents. If the composite value was empty, it was not
possible to tell what the full composite type was. This makes is
impossible to properly do type assertions.

An `Array` and `Map` now need to store the actual parser type within the
value. The simple types can just return the parser "constants" for those
types without needing any extra storage.