parser: Be composite-aware when combining types for literals
Ensure that the inferred type for a composite literal goes as deep as it
can - i.e. `[]any` or `{}any` instead of just `any`. If all values in a
composite type are of the same composite type, then we can infer all
values as a composite of `any` and not just `any`.

The matches the spec where it says:

    The strictest possible type is inferred composite types
    ...
    arr := [ [1] ["a"] ] // type: [][]any

Prior to this change, `arr` would be inferred as `[]any`.