eval: Add typeof builtin, update spec
Add the `typeof` builtin to return the type of a value, either the
type itself of a concrete type, or the underlying type of an `any` type.

Update the spec to specify that a full type is returned for arrays and
maps (rather than just "array" and "map") and that for empty arrays and
maps that the subtype is elided as it is not yet defined.