🐾 go: Use fmt.Sprintf where sensible (#139)
Use fmt.Sprintf where sensible - it was used before with the assumption
that tinygo doesn't support it (runtime error), but it turns out this
was only because we didn't use main() as mount point. While at it,
change all single to double quotes.

Pull-Request: https://github.com/foxygoat/evy/pull/139