🔖 ci: Support auto-versioning of major/minor versions (#145)
Update the `NEXTTAG_CMD` command to identify when a release note has
been added in the latest master commit in `docs/release-notes/v*.md` and
extract the version from the filename. If that version is later than
what `NEXTTAG_CMD` would have used by default, use that version number
for the tag instead.

This allows creating a commit/PR with a new release note and have that
version get auto-tagged and released when merged to master.

Pull-Request: https://github.com/foxygoat/evy/pull/145