docs: Add simple doctest awk script
Add an awk script that formats and runs evy code in markdown files. The
output of the evy code replaces the contents of the associated output
block in the markdown file. Input can be provided to the evy code with
an associated input block.

This can form the basis of a doc testing framework, but for now it is
for ensuring the docs have properly formatted evy and that the output
matches what the program actually outputs.