canvas: Reset to white background, edges with grid
Adjust some of the JS graphics functions, so that a new canvas or a
canvas reset with `clear` or a restart of the program has a `"white"`
rather than a transparent background. If a transparent background is
desired it can still be set with `clear "transparent"`.

I've added this change because it made downloading evy generated images
and using them in the docs easier - I didn't have to remember to first
`clear "white"`. It seems like having a white background is probably
just as desirable as a transparent one, so lets keep it easy for
generating docs for now.

By the same rationale `grid` and `gridn` start at 0 and go all the way
up until 100 now. This outlines the whole canvas, which is desirable
when for examples in Markdown to see where the canvas has its edges.